Replace loader test with standalone version
- Remove original loader.test.ts that had import issues
- Rename loader-standalone.test.ts to loader.test.ts
- Keep the self-contained test that works in CI environments
