--- test-runner.ts
@@ -1,5 +1,0 @@
-#!/usr/bin/env bun
-
-// Simple test runner for command tests
-import "./packages/opencode/test/command/resolver.test.ts"
-import "./packages/opencode/test/command/loader.test.ts"
